+++ examples/index.js
@@ -1,6 +1,2 @@
-const path = require('path');
-
-// change current working directory
-process.chdir(path.join('./', 'examples'));
-
+process.chdir(__dirname);
 require('../');