+++ examples/src/dir1/file1-2.js
@@ -1,2 +1,1 @@
-
 console.log('file1-2');