+++ examples/src/dir2/file2-2.js
@@ -1,2 +1,1 @@
-
 console.log('file2-2');