+++ src/app.js
@@ -1,0 +1,6 @@
+import path from 'path';
+
+const { join } = path;
+
+export const cwd = process.cwd();
+export const tmp = join(cwd, '.md-sync');